# docs/swagger.yaml
basePath: /v1
definitions:
...
  routes.BalanceResponse:
    properties:
      account_id:
        type: string
      as_of:
        description: when projection was computed
        type: string
      available_balance:
        allOf:
        - $ref: '#/definitions/routes.MoneyAmount'
        description: spendable right now
      currency:
        type: string
      held_balance:
        allOf:
        - $ref: '#/definitions/routes.MoneyAmount'
        description: reserved, not yet debited
      ledger_balance:
        allOf:
        - $ref: '#/definitions/routes.MoneyAmount'
        description: total ever posted
    type: object
  routes.DepositeParams:
    properties:
      amount:
        type: string
      client_reference:
        type: string
      currency_code:
        type: string
      customer_id:
        type: string
      description:
        type: string
      destination_account_id:
        type: string
      external_reference:
        type: string
      fee_amount:
        type: string
      idempotency_key_id:
        type: string
      memo:
        type: string
      metadata:
        additionalProperties: true
        type: object
      source_account_id:
        type: string
      source_system:
        type: string
    type: object
...
  routes.MoneyAmount:
    properties:
      amount:
        description: always a string — never float for money
        type: string
      currency:
        description: '"NGN"'
        type: string
    type: object
...
  routes.TransferResponse:
    properties:
      amount:
        $ref: '#/definitions/routes.MoneyAmount'
      client_reference:
        type: string
      description:
        type: string
      destination_account_id:
        type: string
      external_reference:
        type: string
      failure_code:
        type: string
      failure_reason:
        type: string
      fee:
        $ref: '#/definitions/routes.MoneyAmount'
      id:
        type: string
      journal_transaction_id:
        type: string
      posted_at:
        type: string
      requested_at:
        type: string
      source_account_id:
        type: string
      status:
        type: string
    type: object
...
  routes.errResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A comprehensive distributed payment processing system with support
    for transfers, deposits, withdrawals, and holds.
  title: Distributed Payments Engine API
  version: 1.0.0
paths:
  /account/{id}/balances:
    get:
      description: Retrieves all balance information for an account including ledger,
        available, and held amounts.
      parameters:
      - description: Account UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.BalanceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: Get balances for account
      tags:
      - Balances
  /account/deposite:
    post:
      consumes:
      - application/json
      description: Credits funds to an account from the system settlement account.
        Supports idempotent deposits.
      parameters:
      - description: Deposit Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.DepositeParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.TransferResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: Deposit funds to account
      tags:
      - Deposits
...
swagger: "2.0"